\name{biweight_alt_left}
\alias{biweight_alt_left}
\title{Left alternative Biweight kernels}
\usage{
biweight_alt_left(u, rho)
}
\arguments{
\item{u}{Kernel argument.}
\item{rho}{Determines the strength of the boundary adjustment (scalar).}
}
\description{
Left alternative Biweight kernels derived from Müller and Wang (1994, Table 1).
}
\references{
Müller, H.-G. and J.-L. Wang (1994).
``Hazard rate estimation under random censoring with varying kernels and bandwidths''.
\emph{Biometrics} 50 (1), pp. 61--76.
}