\name{epanechnikov_left}
\alias{epanechnikov_left}
\title{Left Epanechnikov kernels}
\usage{
epanechnikov_left(u, rho)
}
\arguments{
\item{u}{Kernel argument.}
\item{rho}{Determines the strength of the boundary adjustment (scalar).}
}
\description{
Left Epanechnikov kernels derived from Müller (1991, Table 1).
}
\references{
Müller, H.-G. (1991).
``Smooth optimum kernel estimators near endpoints''.
\emph{Biometrika} 78 (3), pp. 521--530.
}